test: http2 stream.respond() error checks
PR-URL: https://github.com/nodejs/node/pull/18861
